Добавляет обёртку для тестов
