+++ project/src/components/place-card/place-card.test.tsx
@@ -1,31 +1,13 @@
-import {configureMockStore} from '@jedmao/redux-mock-store';
-import {createMemoryHistory} from 'history';
-import {PlaceCard} from './place-card';
-import {HistoryRouteContainer} from 'components/history-route-container/history-route-container';
 import {makeFakeOffer} from 'utils/mocks/make-fake-offer';
-import {NameSpace} from 'settings/name-space';
-import {Provider} from 'react-redux';
-import {
-  render,
-  screen
-} from '@testing-library/react';
+import {PlaceCard} from './place-card';
+import {testRenderWrapper as render} from 'utils/test-render-wrapper';
+import {screen} from '@testing-library/react';
 
-const mockStore = configureMockStore();
-const history = createMemoryHistory();
 const mockOffer = makeFakeOffer();
-const store = mockStore({
-  [NameSpace.User]: {},
-});
 
 describe('Component: PlaceCard', () => {
   it('should render correctly', () => {
-    render(
-      <Provider store={store}>
-        <HistoryRouteContainer history={history}>
-          <PlaceCard offer={mockOffer}/>
-        </HistoryRouteContainer>
-      </Provider>
-    );
+    render(<PlaceCard offer={mockOffer}/>);
 
     expect(screen.getByTestId('place-card')).toBeInTheDocument();
   });
